RESTの例外ハンドリングのサンプルコードを見直し
