+++ samples/api/api-error-handling/src/main/java/keel/apierrorhandling/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package keel.apierrorhandling;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
 import keel.apierrorhandling.exception.CustomValidationException;
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
@@ -18,7 +17,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
-import java.io.Serializable;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -73,7 +71,7 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
                 ex,
                 messageSource.getMessage(
                         "keel.api-error-handling.HttpMessageNotReadableException",
-                        new Object[0],
+                        null,
                         LocaleContextHolder.getLocale()
                 ),
                 headers,
@@ -113,36 +111,11 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
         return bindingResult
                 .getFieldErrors()
                 .stream()
-                .map(fieldError ->
-                        new ApiError(
-                                fieldError.getField(),
-                                messageSource.getMessage(fieldError, LocaleContextHolder.getLocale())
-                        )
-                )
+                .map(fieldError -> new ApiError(
+                        fieldError.getField(),
+                        messageSource.getMessage(fieldError, LocaleContextHolder.getLocale())))
                 .collect(Collectors.toList());
     }
     // database-validation-end
-
-    static class ApiError implements Serializable {
-
-        @JsonInclude(JsonInclude.Include.NON_EMPTY)
-        private final String field;
-        @JsonInclude(JsonInclude.Include.NON_EMPTY)
-        private final String message;
-
-        ApiError(String field, String message) {
-            this.field = field;
-            this.message = message;
-        }
-
-        public String getField() {
-            return field;
-        }
-
-        public String getMessage() {
-            return message;
-        }
-    }
-
 }
 // api-error-handling-example-end
