+++ samples/api/api-error-handling/src/main/java/keel/apierrorhandling/exception/UserNotFoundException.java
